refactor(podman): extract Installer interface to dedicated file
